# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nexora AI Gateway
  version: 0.1.0-alpha
paths:
  /health:
    get:
      responses:
        "200": {description: Liveness}
  /ready:
    get:
      responses:
        "200": {description: Readiness}
  /v1/models:
    get:
      summary: List active Nexora free models
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Active Nexora models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
  /v1/chat/completions:
    post:
      summary: Create an OpenAI-compatible chat completion
      security: [{bearerAuth: []}]
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: {type: string, maxLength: 128}
          description: Prevents duplicate starts for repeated requests. Same key with a different body returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200": {description: OpenAI-compatible completion}
        "400": {description: Invalid JSON or unsupported parameter}
        "401": {description: Invalid API key}
        "403": {description: Key/model scope not permitted}
        "409": {description: Idempotency conflict}
        "413": {description: Payload too large}
        "422": {description: Unsupported semantic combination}
        "429": {description: Rate limited}
        "502": {description: Unusable upstream response}
        "503": {description: No eligible route available}
        "504": {description: Upstream timeout}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ChatMessage:
      type: object
      required: [role, content]
      additionalProperties: false
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool, developer]
        content:
          description: OpenAI-compatible message content.
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      additionalProperties: false
      properties:
        model:
          type: string
          minLength: 1
          description: Nexora public model ID returned by /v1/models, or auto-free.
        messages:
          type: array
          minItems: 1
          maxItems: 128
          items: {$ref: "#/components/schemas/ChatMessage"}
        temperature: {type: number, minimum: 0, maximum: 2}
        top_p: {type: number, minimum: 0, maximum: 1}
        max_tokens: {type: integer, minimum: 1}
        max_completion_tokens: {type: integer, minimum: 1}
        stream: {type: boolean, default: false}
        stop: {}
        response_format: {}
        tools: {}
        tool_choice: {}
    Model:
      type: object
      required: [id, object, owned_by, free]
      properties:
        id:
          type: string
          description: Provider-neutral Nexora public model ID. Internal routing IDs are never exposed.
        object: {type: string, const: model}
        owned_by: {type: string, const: nexora}
        display_name: {type: string}
        context_length: {type: integer}
        capabilities:
          type: object
          additionalProperties: {type: boolean}
        status: {type: string}
        free: {type: boolean}
    ModelList:
      type: object
      required: [object, data]
      properties:
        object: {type: string, const: list}
        data:
          type: array
          items: {$ref: "#/components/schemas/Model"}